agregado de ventanas de interfaz y ordenamiento de clases
